Renamed preferences components
